requirements.txt
dvc[s3]
scikit-learn
pandas
pyyaml
boto3
Flask
